[test/buildfix/fix] Various fixes to tests
-dfgTestQt:
    -dfgTestCsvItemModel.cpp wouldn't build on Qt6
    -CsvTableView.cpp wouldn't build on some Qt 6 versions (e.g 6.2)
    -Test case dfgQt.TableEditor_fileInfo would fail on older Qt (< 5.10.0)
-dfgTest
    -Silenced a compiler warning in test case DfgUtf.cpToEncoded

+++ dfgTest/dfgTestUtf.cpp
@@ -244,7 +244,7 @@ TEST(DfgUtf, cpToEncoded)
 	// Behaviour in case of invalid cp
 	{
 		std::string s;
-		DFGTEST_EXPECT_LEFT(replacementAscii, cpToEncoded(uint32_max, std::back_inserter(s), encodingLatin1));
+		DFGTEST_EXPECT_LEFT(uint32(replacementAscii), cpToEncoded(uint32_max, std::back_inserter(s), encodingLatin1));
 		DFGTEST_ASSERT_TRUE(s.size() == 1);
 		DFGTEST_EXPECT_LEFT(replacementAscii, s.back());
 		DFGTEST_EXPECT_LEFT(replacementUtf, cpToEncoded(uint32_max, std::back_inserter(s), encodingUTF8));
